+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [17.4.*]
 ### Added
 - SeriesView and ImageViewer classes with text detection and removal using easyocr reader and blackout area with user defined rectangles
-- SeriesView includes whitelist
+- ImageViewer includes histogram display of current frame
+- SeriesView includes whitelist with associated modality specific default lists in assets/locales/*/whitelists
 
 ### Changed
 - Modifications to index and projection modules for memory management and launching SeriesView with ImageViewer